pyproject.toml
[project]
name = "education-mobile"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adrf>=0.1.11",
    "celery[redis]>=5.5.3",
    "colorlog>=6.10.1",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-ratelimit>=4.1.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "drf-spectacular-sidecar>=2025.10.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "ipython>=9.7.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "uvicorn>=0.38.0",
    "uvicorn-worker>=0.4.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
    "ipdb>=0.13.13",
    "whitenoise>=6.11.0",
]
